requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
requests==2.32.3
requests-toolbelt==1.0.0
minio==7.2.20
python-dotenv==1.0.0
gunicorn==23.0.0
google-generativeai==0.8.6